requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
psutil==7.1.0
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
